auth/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
asyncpg==0.30.0
python-dotenv==1.0.1
pydantic==2.10.2
pydantic-settings==2.6.1
msal==1.31.0
pyjwt==2.10.1
httpx==0.28.1
bcrypt==4.2.1
alembic==1.14.0
psycopg2-binary==2.9.10
